pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "zapomnigo python api"
authors = ["Ivan Obreshkov <ivanobreshkov12@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.22"
alembic = "^1.12.1"
flask-migrate = "^4.0.5"
psycopg2-binary = "^2.9.9"
flask-sqlalchemy = "^3.1.1"
flask-bcrypt = "^1.0.1"
python-ulid = "^2.2.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
pyjwt = "^2.8.0"
flask-cors = "^4.0.0"
asgiref = "^3.7.2"
flask-limiter = "^3.5.0"
lxml = "^5.1.0"
celery = "^5.3.6"
redis = "^5.0.2"
gunicorn = "^21.2.0"

[tool.poetry.scripts]
start = "src.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
